refactor: clean code function, code stell working.
