--- src/satellite/orbcomm/helpers/func/upload-functions.ts
@@ -1,58 +1,0 @@
-import { HttpService } from '@nestjs/axios';
-import {
-  ForwardStatuses,
-  MessageBodyGetStatus,
-  MessageBodyPost,
-  StatusesType,
-  Submission,
-} from '../interfaces/upload-messages.interfaces';
-
-export const postApiMessages = (http: HttpService, body: MessageBodyPost) => {
-  return new Promise<Submission[]>((resolve, reject) => {
-    http.axiosRef
-      .post('http://localhost:3001/orbcomm/post', {
-        body,
-      })
-      .then((apiResponse) => {
-        console.log(apiResponse.data);
-        const correctValues = verifyPostMessages(
-          body.messages,
-          apiResponse.data,
-        );
-        console.log(correctValues);
-        resolve(correctValues);
-      })
-      .catch((error) => {
-        reject(error.message);
-      });
-  });
-};
-
-export const verifyPostMessages = (sendedData, responseData): Submission[] => {
-  const validItems: Submission[] = [];
-  responseData.Submissions.map((apiResponse) => {
-    const exists = sendedData.find(
-      (data) => data.UserMessageID === apiResponse.UserMessageID,
-    );
-    if (exists) {
-      validItems.push(apiResponse);
-    }
-  });
-  return validItems;
-};
-
-export const getMessagesOrbcommStatus = (
-  http: HttpService,
-  body: MessageBodyGetStatus,
-) => {
-  return new Promise<ForwardStatuses>((resolve, reject) => {
-    http.axiosRef
-      .get('http://localhost:3001/orbcomm/getfwd', { params: body })
-      .then((apiResponse) => {
-        resolve(apiResponse.data);
-      })
-      .catch((error) => {
-        reject(error.message);
-      });
-  });
-};
