+++ src/satellite/orbcomm/helpers/index.ts
@@ -1,4 +1,6 @@
-export * from './dtos/upload-message.dto';
-export * from './func/upload-functions';
-export * from './interfaces/upload-messages.interfaces';
-export * from './validators/orbcomm.validators';
+export * from './dtos/orbcomm.dto';
+export * from './func/http/orbcomm-http.func';
+export * from './interfaces/orbcomm-messages.interfaces';
+export * from './func/validators/orbcomm-validators.func';
+export * from './func/format/orbcomm-format.func';
+export * from './func/db/orbcomm-db.func';
