+++ src/satellite/orbcomm/helpers/interfaces/orbcomm-messages.interfaces.ts
